支持浮点

+++ src/main/java/com/mao/rocket/controller/mvc/BillResponse.java
@@ -7,7 +7,7 @@ import java.util.List;
 public class BillResponse {
   public long id;
   public String userName;
-  public long amount;
+  public double amount;
   public String status;
   public List<Item> items;
 }
