+++ src/main/java/com/mao/rocket/model/vo/Item.java
@@ -3,7 +3,7 @@ package com.mao.rocket.model.vo;
 public class Item {
   public long id;
   public String name;
-  public long price;
+  public double price;
   public String image;
 
   public long getId() {
@@ -22,11 +22,11 @@ public class Item {
     this.name = name;
   }
 
-  public long getPrice() {
+  public double getPrice() {
     return price;
   }
 
-  public void setPrice(long price) {
+  public void setPrice(double price) {
     this.price = price;
   }
